pyproject.toml
[tool.poetry]
name = "buppy"
version = "0.1.0"
description = ""
authors = ["Youngwook Kim <youngwook.kim@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8.6,<3.9.7"
langchain = {extras = ["llms"], version = "^0.0.332"}
slack-bolt = "^1.18.0"
emoji-data-python = "^1.5.0"
google-cloud-firestore = "^2.13.1"
aiohttp = "^3.8.6"
streamlit = "^1.28.2"
apscheduler = "^3.10.4"
pydantic-settings = "^2.1.0"
pydantic = "^2.5.2"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pylint = "^2.17.4"
mypy = "^1.4.1"
pyright = "^1.1.316"
isort = "^5.12.0"
omgpt = "0.0.7"
watchdog = "^3.0.0"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
